+++ src/main/java/de/fhswf/informatik/se/praktikum8/notenverwaltung/backend/service/WahlmodulService.java
@@ -30,6 +30,9 @@ public class WahlmodulService {
     }
 
     public void deleteWahlmodul(String modulname){
+        if(repository.findByModulname(modulname) == null){
+            throw new IllegalArgumentException("Das gewählte Wahlmodul wurde nicht gefunden.");
+        }
         Wahlmodul wahlmodul = repository.findByModulname(modulname);
         repository.delete(wahlmodul);
     }
